Update supported-platforms.md
In this Supported Platforms page, we've mentioned that RHEL 7.7 is supported but we did not specify that it is supported starting with 5.6 and later.  In the dropdown button to select the version, we only see "5.6 (latest)" and "5.2 (and below)".  Users who are using 5.4 would select the 5.6 (latest) and would assume that RHEL 7.7 is now supported in 5.4 as well.