Update crdb-cli-reference.md
Updating the memory size and shards count are the common updates in CRDB.  However, we did not mention the `default-db-config` property and that sub properties can be specified in this doc.  We've mentioned the `db-config-json` but no details on what the content should be like.   In the unofficial doc and `crdb-cli crdb update --help`, we mentioned the use of `default-db-config` while the public doc mentioned `db-config-json`.